+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Backstabs
 
+## 2.3
+
+### Features
+
+- FOV check is ignored if the player has 100% or more chameleon effect
+- Fov check is ignored if the player has invisibility effect
+
 ## 2.2.1
 
 ### Features